requirements.txt
argparse>=1.2.1
coverage>=3.7.1
distribute>=0.6.24
leap.common>=0.3.8
leap.soledad.client>=0.5.2-22-g4519775
leap.soledad.common>=0.5.2-22-g4519775
mock>=1.0.1
protobuf>=2.5.0
pyzmq>=14.3.1
simplejson>=3.5.3
u1db>=13.09
wsgiref>=0.1.2
